changed token attributes to long long int and double

+++ scanner.h
@@ -15,8 +15,8 @@ IFJ project
 #include "enums.h"
 
 typedef union {
-    int integer;
-    float decimal;
+    long long int integer;
+    double decimal;
     char *string;
     Keyword keyword;
 } TokenAttribute;
